+++ src/global/Button.css
@@ -1,0 +1,27 @@
+.button {
+	composes: surface from '../static/styles.css';
+	padding: 0.25em 1em 0.4em;
+	text-align: center;
+	display: inline-block;
+	font-size: 1em;
+}
+.button:active {
+	border-bottom-color: rgba(255, 255, 255, 0.75);
+	border-right-color: rgba(255, 255, 255, 0.75);
+	border-left-color: #000;
+	border-top-color: #000;
+}
+.button:active:before {
+	border-top-color: rgba(0, 0, 0, 0.25);
+	border-left-color: rgba(0, 0, 0, 0.25);
+	border-right-color: rgba(255, 255, 255, 0.25);
+	border-bottom-color: rgba(255, 255, 255, 0.25);
+}
+
+.button:focus {
+	outline: none;
+}
+.button:focus:after {
+	border: 1px dotted #000;
+	margin: 1px 2px 2px 1px;
+}
