+++ src/global/WaveWindow.css
@@ -1,0 +1,97 @@
+.wrapper {
+	position: fixed;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	overflow: scroll;
+	display: flex;
+	align-content: center;
+	align-items: center;
+	justify-content: center;
+	padding-bottom: 4em;
+	pointer-events: none;
+	font-size: 11px;
+	letter-spacing: 0.33px;
+	-webkit-font-smoothing: none;
+	font-family: tahoma, sans-serif;
+}
+.wrapper[data-dark='true'] {
+	background: url(data:image/gif;base64,R0lGODlhAgACAIAAAP///wAAACH5BAEAAAAALAAAAAACAAIAQAIDDBAFADs=);
+	pointer-events: all;
+}
+.wrapper * {
+	pointer-events: all;
+}
+
+.window {
+	composes: surface from '../static/styles.css';
+	padding: 2px;
+	border: 1px solid rgba(0, 0, 0, 1);
+	border-top: none;
+	border-left: none;
+	max-width: 300px;
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	align-items: stretch;
+	align-content: stretch;
+	font-kerning: none;
+}
+.window:before {
+	border-top-color: rgba(255, 255, 255, 0.75);
+	border-left-color: rgba(255, 255, 255, 0.75);
+}
+.windowTitle {
+	flex: 0 0 auto;
+	min-height: 1px;
+}
+.window:focus .windowTitle,
+.window:focus-within .windowTitle {
+	background-color: #000080;
+	color: #fff;
+}
+.window:focus {
+	outline: none;
+}
+
+.windowTitle {
+	padding: 0.25em 0.25em 0.25em 0.5em;
+	height: 1.2em;
+	font-weight: 800;
+	background-image: linear-gradient(
+			to right,
+			rgba(255, 255, 255, 0),
+			rgba(255, 255, 255, 1)
+		),
+		linear-gradient(to right, rgba(255, 255, 255, 0), rgba(255, 255, 255, 0.2));
+	background-blend-mode: overlay, lighten;
+}
+
+.titleButton {
+	composes: surface from '../static/styles.css';
+	float: right;
+	height: 1.25em;
+	width: 1.25em;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding: 0;
+	color: #000;
+	text-align: center;
+	font-weight: 800;
+	line-height: 0.9em;
+}
+
+.titleButton:active {
+	border-bottom-color: rgba(255, 255, 255, 0.75);
+	border-right-color: rgba(255, 255, 255, 0.75);
+	border-left-color: #000;
+	border-top-color: #000;
+}
+.titleButton:active:before {
+	border-top-color: rgba(0, 0, 0, 0.25);
+	border-left-color: rgba(0, 0, 0, 0.25);
+	border-right-color: rgba(255, 255, 255, 0.25);
+	border-bottom-color: rgba(255, 255, 255, 0.25);
+}
