+++ src/global/dialog.css
@@ -1,5 +1,5 @@
 .dialog {
-	padding: 1em;
+	padding: 0.75em;
 }
 
 .message {
@@ -16,7 +16,7 @@
 
 .buttons {
 	display: grid;
-	grid-template-columns: 5em;
+	grid-auto-columns: 6em;
 	grid-auto-flow: column;
 	grid-gap: 0.25em;
 	justify-content: center;
