+++ src/global/dialog.js
@@ -1,29 +1,22 @@
 import { $element } from '../util';
 
-import { WaveWindow } from './wavewindow';
-import { message, buttons, dialog } from './dialog.css';
+import { message, buttons as buttonsCss, dialog } from './Dialog.css';
 import { glyph } from '../static/glyph.css';
 
-export class Dialog extends WaveWindow {
-	get buttons() {
-		return [];
+export const Dialog = ({ buttons, icon, text }) => {
+	if (!text) {
+		text = 'The window at ' + window.location.href + ' says nothing';
 	}
-
-	get clientArea() {
-		const $buttons = $element('div', { class: buttons }, this.buttons);
-		const $message = $element(
-			'div',
-			{ class: message, 'data-icon': !!this.params.icon },
-			[
-				this.params.icon &&
-					$element('span', {
-						class: glyph,
-						'data-glyph': this.params.icon,
-					}),
-				this.params.text,
-			]
-		);
-		const $dialog = $element('div', { class: dialog }, [$message, $buttons]);
-		return $dialog;
-	}
-}
+	const $dialog = $element('div', { class: dialog }, [
+		$element('div', { class: message, 'data-icon': !!icon }, [
+			icon &&
+				$element('span', {
+					class: glyph,
+					'data-glyph': icon,
+				}),
+			text,
+		]),
+		$element('div', { class: buttonsCss }, buttons),
+	]);
+	return $dialog;
+};
