+++ src/static/styles.css
@@ -11,10 +11,6 @@
 	text-decoration: inherit;
 	padding: 1px;
 	position: relative;
-	font: 1em tahoma, verdana, sans-serif;
-	letter-spacing: 0.5px;
-	-webkit-font-smoothing: none;
-	font-smoothing: none;
 	border: 1px solid #000;
 	border-top-color: rgba(255, 255, 255, 0.75);
 	border-left-color: rgba(255, 255, 255, 0.75);
@@ -42,118 +38,4 @@
 }
 .surface * {
 	z-index: 10;
-}
-
-.wrapper {
-	position: fixed;
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	overflow: scroll;
-	display: flex;
-	align-content: center;
-	align-items: center;
-	justify-content: center;
-	padding-bottom: 4em;
-	pointer-events: none;
-}
-.wrapper[data-dark='true'] {
-	background: url(data:image/gif;base64,R0lGODlhAgACAIAAAP///wAAACH5BAEAAAAALAAAAAACAAIAQAIDDBAFADs=);
-	pointer-events: all;
-}
-.wrapper * {
-	pointer-events: all;
-}
-
-.button {
-	composes: surface;
-	padding: 0.25em 1em 0.4em;
-	text-align: center;
-	display: inline-block;
-}
-.button:active {
-	border-bottom-color: rgba(255, 255, 255, 0.75);
-	border-right-color: rgba(255, 255, 255, 0.75);
-	border-left-color: #000;
-	border-top-color: #000;
-}
-.button:active:before {
-	border-top-color: rgba(0, 0, 0, 0.25);
-	border-left-color: rgba(0, 0, 0, 0.25);
-	border-right-color: rgba(255, 255, 255, 0.25);
-	border-bottom-color: rgba(255, 255, 255, 0.25);
-}
-
-.button:focus {
-	outline: none;
-}
-.button:focus:after {
-	border: 1px dotted #000;
-	margin: 1px 2px 2px 1px;
-}
-
-.window {
-	composes: surface;
-	padding: 2px;
-	border: 1px solid rgba(0, 0, 0, 1);
-	border-top: none;
-	border-left: none;
-	max-width: 300px;
-	font-size: 0.7em;
-	position: relative;
-	display: flex;
-	flex-direction: column;
-	align-items: stretch;
-	align-content: stretch;
-}
-.window:before {
-	border-top-color: rgba(255, 255, 255, 0.75);
-	border-left-color: rgba(255, 255, 255, 0.75);
-}
-.windowTitle {
-	flex: 0 0 auto;
-	min-height: 1px;
-}
-.window:focus .windowTitle,
-.window:focus-within .windowTitle {
-	background: #000080;
-	color: #fff;
-}
-.window .clientArea {
-	flex: 1 0 auto;
-	min-height: 1px;
-}
-
-.clientArea {
-	position: relative;
-}
-
-.content {
-	padding: 0.5em;
-	margin: 0.35em 0.15em 0.15em 0.15em;
-}
-
-.windowTitle {
-	padding: 0.25em 0.25em 0.25em 0.5em;
-	height: 1.2em;
-	font-weight: 800;
-	background-image: linear-gradient(
-			to right,
-			rgba(255, 255, 255, 0),
-			rgba(255, 255, 255, 1)
-		),
-		linear-gradient(to right, rgba(255, 255, 255, 0), rgba(255, 255, 255, 0.2));
-	background-blend-mode: overlay, lighten;
-}
-.titleButton {
-	composes: surface;
-	float: right;
-	height: 1.1em;
-	width: 1.25em;
-	padding: 0;
-	color: #000;
-	text-align: center;
-	font-weight: 800;
-	line-height: 0.9em;
 }
